% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqverse-package.R
\docType{package}
\name{pqverse-package}
\alias{pqverse}
\alias{pqverse-package}
\title{pqverse: Easily Install and Load the 'pqverse'}
\description{
The 'pqverse' is a set of packages for metabarcoding analysis that share the phyloseq object as their central data structure. This package makes it easy to install and load multiple 'pqverse' packages in a single step.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adrientaudiere/pqverse}
  \item Report bugs at \url{https://github.com/adrientaudiere/pqverse/issues}
}

}
\author{
\strong{Maintainer}: Adrien Taudière \email{adrien.taudiere@zaclys.net} (\href{https://orcid.org/0000-0003-1088-1182}{ORCID})

}
\keyword{internal}
