% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{pqverse_packages}
\alias{pqverse_packages}
\title{List all pqverse packages}
\usage{
pqverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include pqverse in the list?}
}
\value{
A character vector of package names.
}
\description{
List all pqverse packages
}
